+++ contracts/interfaces/IBalanceAdder.sol
@@ -2,5 +2,5 @@
 pragma solidity >=0.8.0;
 
 interface IBalanceAdder {
-    function processBalances(uint step) external;
+    function processBalances(uint256 step) external;
 }
